+++ .env.example
@@ -11,6 +11,7 @@ PORT=5000
 MONGO_URI=mongodb://root:root@localhost:27017/training?authSource=admin
 
 # Cookie
+# Should add secure on production.
 COOKIE_OPTIONS=SameSite=Strict; HttpOnly;
 
 # Bcrypt
