+++ src/controllers/UserController.js
@@ -1,4 +1,4 @@
-const User = require('./user.model')
+const User = require('../models/User')
 
 module.exports = {
     async save({ body }, res) {
