+++ src/middlewares/UserMiddleware.js
@@ -1,4 +1,4 @@
-const User = require('./user.model')
+const User = require('../models/User')
 
 module.exports = {
     async userExists({ body }, res, next) {
